More calls to normalizePath
The `services` map in MicroservicesRegistryImpl uses resource paths as key.
Therefore, we should normalize these path so that we e.g. treat "a//b" and "a/b"
and "/a/b" as equal.

A better solution might be to create a "ResourcePath" type and hide all that
normalization within it, but that would be a more intrusive change.